[Feat] 몬스터 체력바 연동 및 크기 설정

+++ Assets/1.Worker/WGH/Scripts/GameProgress.cs
@@ -17,7 +17,7 @@ public class GameProgress
 {
     public int curStage;
 
-    public float playerDmg;
+    public long playerDmg;
     public float criticalChance;
     public float criticalDmgPercent;
